Moving generic submission grid deserialization to a common place. Adding it to hidden, container, and custom components.
