+++ src/components/container.js
@@ -1,4 +1,6 @@
 var fs = require('fs');
+var GridUtils = require('../factories/GridUtils')();
+
 module.exports = function(app) {
   app.config([
     'formioComponentsProvider',
@@ -9,41 +11,7 @@ module.exports = function(app) {
         viewTemplate: 'formio/componentsView/container.html',
         group: 'advanced',
         icon: 'fa fa-folder-open',
-        tableView: function(data, component, $interpolate, componentInfo) {
-          var view = '<table class="table table-striped table-bordered"><thead><tr>';
-          angular.forEach(component.components, function(component) {
-            view += '<th>' + component.label + '</th>';
-          });
-          view += '</tr></thead>';
-          view += '<tbody>';
-          view += '<tr>';
-          angular.forEach(component.components, function(component) {
-            var info = componentInfo.components.hasOwnProperty(component.type) ? componentInfo.components[component.type] : {};
-            if (info.tableView) {
-              view += '<td>' +
-                info.tableView(
-                  data && component.key && data.hasOwnProperty(component.key) ? data[component.key] : '',
-                  component,
-                  $interpolate,
-                  componentInfo
-                ) + '</td>';
-              return;
-            }
-
-            view += '<td>';
-            if (component.prefix) {
-              view += component.prefix;
-            }
-            view += data[component.key] || '';
-            if (component.suffix) {
-              view += ' ' + component.suffix;
-            }
-            view += '</td>';
-          });
-          view += '</tr>';
-          view += '</tbody></table>';
-          return view;
-        },
+        tableView: GridUtils.generic,
         settings: {
           input: true,
           tree: true,
